Rename generated and base client class package
