+++ ygoprodeck-client/src/test/java/io/github/magicdgs/gaming/ygoprodeck/client/MockYgoprodeckRetrofitSyncClientTest.java
@@ -1,0 +1,8 @@
+package io.github.magicdgs.gaming.ygoprodeck.client;
+
+public class MockYgoprodeckRetrofitSyncClientTest extends AbstractMockYgoprodeckClientTest {
+    @Override
+    protected boolean isAsync() {
+        return false;
+    }
+}
