--- ygoprodeck-client/src/test/java/io/github/magicdgs/gaming/ygoprodeck/client/retrofit/MockYgoprodeckRetrofitAsyncClientTest.java
@@ -1,8 +1,0 @@
-package io.github.magicdgs.gaming.ygoprodeck.client.retrofit;
-
-public class MockYgoprodeckRetrofitAsyncClientTest extends AbstractMockYgoprodeckRetrofitClientTest {
-    @Override
-    protected boolean isAsync() {
-        return true;
-    }
-}
