--- ygoprodeck-client/src/test/java/io/github/magicdgs/gaming/ygoprodeck/client/retrofit/MockYgoprodeckRetrofitSyncClientTest.java
@@ -1,8 +1,0 @@
-package io.github.magicdgs.gaming.ygoprodeck.client.retrofit;
-
-public class MockYgoprodeckRetrofitSyncClientTest extends AbstractMockYgoprodeckRetrofitClientTest {
-    @Override
-    protected boolean isAsync() {
-        return false;
-    }
-}
